docs(elements): move SingleTag to elements
