+++ src/index.ts
@@ -23,7 +23,7 @@ export { Dropdown } from './components/Dropdown'
 export { NewWindow } from './components/NewWindow'
 export { NotificationEvent, Notifier } from './components/Notifier'
 export { SideMenu } from './components/SideMenu'
-export { SingleTag } from './components/SingleTag'
+export { SingleTag } from './elements/SingleTag'
 export { MapMenuDialog } from './components/MapMenuDialog'
 export { Message } from './components/Message'
 
@@ -141,7 +141,7 @@ export type { DropdownProps, DropdownItemProps } from './components/Dropdown'
 export type { NewWindowProps } from './components/NewWindow'
 export type { NotifierProps } from './components/Notifier'
 export type { SideMenuProps } from './components/SideMenu'
-export type { SingleTagProps } from './components/SingleTag'
+export type { SingleTagProps } from './elements/SingleTag'
 export type { MessageProps } from './components/Message'
 
 export type { ButtonProps } from './elements/Button'
